Remove hard-hyphen handling :(
Joining word fragments around hard hyphens can result in
mismatches due to hard hyphens used in lists (for example: "the
Syrens- or Fiſh-men"). This kind of sentence occurs a lot in
Cavendish's The Blazing World, resulting in so-called complete
words like "Syrens-or".

Since hard hyphens are generally used to connect whole words
anyway, it should be fine to not resolve them.