Propagate @read to descendant elements as necessary
If an element should be have its text content hidden via @read,
the XML tree is no longer flattened when that element had a child
element or both element and text children. Instead, @reads are
placed on the descendant elements, and wrapper <seg>s are created
for text nodes that are part of mixed content.

Intervention @type/@subtypes are more explicit about what actions
are taken on the content of an element, and when an element is
itself created.